+++ src/apps/sync-engine/dependency-injection/files/builder.ts
@@ -1,5 +1,5 @@
 import crypt from '../../../../context/shared/infrastructure/crypt';
-import { CreateFilePlaceholderOnDeletionFailed } from '../../../../context/virtual-drive/files/application/CreateFilePlaceholderOnDeletionFailed';
+import { CreateFilePlaceholderOnDeletionFailed } from '../../../../context/virtual-drive/files/application/event-subsribers/CreateFilePlaceholderOnDeletionFailed';
 import { FileCreator } from '../../../../context/virtual-drive/files/application/FileCreator';
 import { FileDeleter } from '../../../../context/virtual-drive/files/application/FileDeleter';
 import { FilePathUpdater } from '../../../../context/virtual-drive/files/application/FilePathUpdater';
