+++ src/context/virtual-drive/files/application/event-subsribers/CreateFileOnOfflineFileUplodaded.ts
@@ -1,10 +1,10 @@
-import { OfflineContentsUploadedDomainEvent } from '../../../offline-drive/contents/domain/events/OfflineContentsUploadedDomainEvent';
-import { DomainEventClass } from '../../../shared/domain/DomainEvent';
-import { DomainEventSubscriber } from '../../../shared/domain/DomainEventSubscriber';
-import { FilePath } from '../domain/FilePath';
-import { FileCreator } from './FileCreator';
+import { OfflineContentsUploadedDomainEvent } from '../../../../offline-drive/contents/domain/events/OfflineContentsUploadedDomainEvent';
+import { DomainEventClass } from '../../../../shared/domain/DomainEvent';
+import { DomainEventSubscriber } from '../../../../shared/domain/DomainEventSubscriber';
+import { FilePath } from '../../domain/FilePath';
+import { FileCreator } from '../FileCreator';
 import Logger from 'electron-log';
-import { FileToOverrideProvider } from './FileToOverrideProvider';
+import { FileToOverrideProvider } from '../FileToOverrideProvider';
 
 export class CreateFileOnOfflineFileUploaded
   implements DomainEventSubscriber<OfflineContentsUploadedDomainEvent>
